Bump TagController
Fixes only zeroing out half a line when promoting a non-leaf 0 to a 1,
causing the upper half of some pages to have junk in the tag bits when
storing the first tagged quantity there after reset. Also includes
several other cleanups and fixes for bugs that happen to not be
triggered by Flute as things stand.